Refactored API to use Optional<AEItemDefinition>
The new definitions can now be access through API.definitions(),
containing methods to access blocks, items, materials and parts as before.

AEItemDefinitions can now be checked by Optional.isPresent() instead of
handling null and possible NPEs.

The old methods in API to access blocks, items, materials and parts are
now deprecated and scheduled for removal.

Addon developers are free to ignore it for rv2 to prevent backward
incompatibilities. But it is highly recommended to get familiar with it
and migrate with the next revision.