Introduction to the new Optional API
Deprecates the old usage of the AEItemDefinitions via the direct method access of

* blocks()
* parts()
* items()
* materials()

and thus use the new re-direct via definitions().

The whole change is just a temporary major fix to the nagging of the missing definitions
due to the possibility that they being null if being deactivated by config.

For future reference, it would be better to initialize the block and item definitions,
but intercept the registration, so the user cant access it, but the processing,
will not fail because the definitions will still be valid. This could be a API change introduced in Minecraft 1.8+