Updated the buildscript and dependencies
It now uses a single maven proxy, but the old repos are still kept as
documentation and new ones should also be documented.
The old ones can potentially be uncommented, it is mostly a question of
how fast the lookups are with multiple repos compared to a single proxy.

Self hosted artifacts or self compiled APIs are now using the actual mod
version they are based on for a better documentation in case of any api
breaking change made by them.

Splitted dependencies and repositories into two files as they are getting
a bit large.

Updating immibis contains a renamed Interface, this is also fixed.