Improve behaviour of stacks that have 0 stackSize and non-zero requestable count
Without these changes such stacks appear as craftable in the ME Terminal, but are unable to proceed past the Craft Confirm GUI, and also unable to be used in autocrafting.

With these changes the ME Terminal will not display Craft text when requestable count > 0. Stacks with 0 stack size and requestable count > 0 will display no number in the slot. Crafting calculation can successfully use the requestable stacks as if they existed.