Fixes a ItemStack deletion issue in the grindstone
This fixes the Grinder eating the item in case of an input
ItemStack with size equal to one.
The important bit here is the sequence of instructions.
With the <=1.10 representation of empty ItemStacks using null
the old ordering was fine. However, whenever an ItemStack gets
shrinked to zero, its internal Item gets replaced, thus you
have to make a copy before shrinking.

Additionally using ItemStack::shrink instead of grow with
negative value.