Refactored the BaseActionSource
It now uses an interface instead of a concrete class and further
subclasses.

Instead of relying on a specific class for a certain action type, it now
uses methods with Optional as return values to determine a player or
machine issuing an action. Refer to the JavaDocs for the exact behaviour.

IActionHost no longer extends IGridHost. It never used the additional
functionality and if needed the IGridNode will also provide a reference
to the corresponding IGridHost.