Use @Instance for mod lifecycle.
This is janitorial work to bring things up to date with modern Forge
practice. There's no need for a factory, just annotate a public data
member and Forge promises to manage the mod as a singleton.

The factory method gets removed, too. We don't need it since Forge is
fine just calling the (now public) no-arg constructor.