(#2842) Re-add support for the OpenComputers network P2P Tunnel
The OpenComputers P2P tunnel was disabled during the port to 1.8. This
commit re-adds the OpenComputers P2P tunnel. The implementation is based
on the OpenComputers P2P tunnel code from AE2 in 1.7.10, updated and
refactored based on the IC2 P2P tunnel and general integration module
design.

Tested and working on Minecraft 1.12.2, Forge version 14.23.4.2750,
with OpenComputers version 1.7.2.67, connecting a RAID block, tier 3
screen and keyboard to both a tier 3 computer and a tier 3 server.
Also successfully tested AE2 with the OpenComputers mod removed.