(#3854) Place properly colored cables when holding Color Applicator in the other hand
Currently placing cables while also holding a Color Applicator does not place cables colored per the Color Applicator. With this fix, when placing cables while also holding a Color Applicator the cables will be placed with the proper color and dye will be removed from the Color Applicator. This works with either the cable or Color applicator in the off hand.