(#3800) Don't fire PowerStatusChange events while extracting power.
EnergyGridCache.extractProviderPower holds an iterator over its provider
set while calling extractAEPower on each provider.  When it extracts
power from its localStorage and empties it, firing a PowerStatusChange
event immediately can de-power a TileController and trigger the
destruction of the ME grid which both the TileController and
EnergyGridCache are part of. This in turn causes the TileController to
be removed from the EnergyGridCache's provider set while the iterator is
still live, resulting in a ConcurrentModificationException.

Evidence in this backtrace: https://pastebin.com/yJLtAcx8

Instead of firing the event immediately, set a semaphore to notify the
EnergyGridCache the localStorage is out of power and rely on the
already-existing logic to fire the PowerStatusChange once grid power
extraction has completed.