Fixes #4119 Sort the list of items so that missing items appear first
So I'm not sure if this is in the right style or whatever, so happy to take suggestions. I do think this approach (i.e. sorting explicitly in the GUI code) is better than the current approach which relies on the fact that packets are received in a certain order to put things in the right order.

Here we explicitly sort the items so that missing items appear first. Secondarily, items with the biggest stack appear second.