Kill CompassService when server is closed, and not closing.
This is done in order to ensure that specified TEs, specifically CraftingTileEntity do not submit requests to the CompassService after it has already been killed. This fixes #4890

If the better way to fix this issue is to instead make sure that no TEs submit requests after the CompassService has closed, then I am unable to help with that.