Added rest of outline's Categories.  Updated Patchouli version.
Updated patchouli from 1.16.2-48-SNAPSHOT to 1.16.4-48-SNAPSHOT.  This is necessary in order to have more than 4 sub-cagetories on a page (when there are no entries in the containing Category).

Updated book.json to add more formatting options.  I am using the new background texture (as a resource pack), but that's not my work to upload.  Removing the paper page turning sounds using null values is an enchancement that Patchouli will be adding at some point.

Added the rest of the Categoris and Sub-Categories listed in the outline.  I've added a Misc. category because I needed somewhere to put a "Channels" entry.  Updated the existing categories to use appliedenergistic2's namespace since I now have the dev environment working.