Adds a network wide chunkloader (#4998)
Textures by @Ridanisaurus

This makes use of the readded `ForgeChunkManager` making it much more easy to restore the chunk loaders during load as well as also easily handling ones removed by external means. (Backups, nbt editing, etc).
**However this has the limitation that we can only load their own world/dimension.** Networks spanning multipe worlds via QNBs or other means need their own chunk loader. Otherwise we would have to implement handling the chunk tickets completely on your side.

Adds a new `MENetworkStatisticsEvent` with subtypes related to the grid entering or leaving new chunks. This could be further extended in the future for other use cases like debugging.
**For now the new `IGridCache` providing these informations is not part of the public API and should not be used until changed**

It adds a new block, the corresponding tile entity, container and screen for it.
The UI is not fully polished yet and might need some work, but provides details about the energy consumption of this chunk loader, how many chunks it loads as well as how many chunks and worlds the whole network spans as well as how many of them are loaded. This only takes our chunk loaders into account.

The UI also provides a toggle to display an in world overlay marking all loaded chunks of this network.

Power consumption per tick is not fully finalised and needs some evaluation. The basic idea is to have it increasingly scale with the amount of chunks and worlds. Probably with a minimum value aimed towards midgame, but certainly not endgame. This should be covered by the scaling.

There is currently no way to disable its behaviour and most likely not planned. If this is not intended, it is recommended to disable the recipe and use whatever forge otherwise provides to disable it or prevents player from acquiring or using already existing ones.

Resolves #4966