Downgraded JEI missing slots error to a warning.
Now using the new JEI feature added in 7.7.0.98 to indicate a warning by
tinting the `+` button yellow/orange instead of a hard fail.
Update to using the new methods using `ITextCompoment` instead of
`String`
Fixed a wrong text entry for charged certus quartz when the worldgen is
disabled.

Updated to JEI 7.7.0.98 as minimum requirement
Updated to Forge 36.1.10 as minimum requirement to fix it ignoring
minimum requirements for non mandatory dependencies.