Remove tunnel type, allow for addon P2P tunnels. (#5622)
By removing TunnelType and switching to a system where the tunnel type is communicated via the part item, addons should now be able to add new P2P tunnels that work like the built-in ones in terms of attunement.

Addons will still need to depend on internals for the actual tunnel item.