Change ActionKey.Toggle_Focus Key
This commit changes the hardcoded ActionKey.Toggle_Focus keybind from Tab to Equals sign (=). This alleviates a conflict with assigning the Tab key to open/close inventory which is a popular binding. Equals sign is much less frequently used. This fix is temporary until a client changeable keybind for Toggle_Focus is added.