Implement missingIngredients
`continue entityLoop` still required if a recipe uses an item more than once, so it remains intact behind the "does the recipe still needs this item" check.

as for the `!missingIngredients.isEmpty()` check, does it really matter if there is already a `selectedEntities.size() == recipe.ingredients.size()` check in there? i guess it could be replaced instead of substituted.