Fix issues relating to Annihilation Plane checking if it can store a block and then breaking it in two separate ticks (by doing it all in one tick)
The problem was most obvious when you had a setup with X annihilation planes and 1 formation plane as the only storage; if multiple planes tried to pickup a block in the same tick, one will store it and the rest will drop the item on the ground, as they all succeeded on the simulated insert, but only one will be able to actually insert on the next tick. But the issue could be seen any time the networks ability to store the break result changes between the first and second break process tick (or if two loot table rolls gave different results, and only the first could actually be stored).