Fix crafting job progress bar exceeding 100%
This fixes a progress bar rendering more than the maximum width by
clamping the percentage if the crafting job makes heavy use of
non-consumed items.

Note that this is merely a cosmetic fix for an overly long progress bar.
The bar will still reach 100% much faster than completing the job. This
is a more fundamental issue that still needs to be fixed separately.

Partial fix for #7468.