Fix crafting job progress bar exceeding 100%
Fix for 1.21 for the bug reported independently in #7468 for 1.20
and #7420 for 1.19.

The bug stems from the discrepancy between the initial calculation that
only accounts for recipe outputs and the actual execution of the job
which adds recipe outputs as well as any remaining items in the crafting
grid, called container items. When decrementing the counter for the
remaining items, these container items must not be counted to ensure the
counter only decrements down to 0 and not beyond.

Fixing this side of the problem, rather than the initial calculation has
two reasons:
  - more accurate progress bar, because recipes with container items
    don't decrease the counter multiple times per recipe execution, which
    will become important for more complex crafting jobs with a mix of normal
    recipes and those with additional container items.
  - the actual number of container items is determined only at the
    execution stage and not during the initial calculation.