Fix resuming crafting recipe jobs after a world load
The ItemStack type used in EncodedCraftingPattern provides no
hashCode() function and is thus not usable in hashCode() functions
of objects containing it.

This leads to an effect where a running job using an
EncodedCraftingPattern does not resume after loading the world.
Providing an explicit hashCode() function for EncodedCraftingPattern
solves this problem.

Fix #8006.