(#8274) Optimize PathingService and GridConnection repath conditions
There are several ways that a network can enter the "rebooting" state
however only a few actually have consequence to the overall network
state.

This change removes the need for the network to reboot during
"simple" Grid merges while still enforcing rebooting for "complex"
merges. A simple merge is one that has no bearing on the network
pathing; imagine a long cable with no channel-consuming devices attached
to it and then you add another cable. A complex network merge is
everything else. Of note: half-items like P2P, Quartz Fiber, and all
terminals always force a complex merge for some reason but seeing as the
network needs to reboot to calculate their energy/channel consumption;
this is preferred.