Fix #8329: StorageService updates getting more expensive over time (#8331)
Fixes #8329 which was caused by the inner maps of
`cachedAvailableStacks` never getting removed: once the primary key of
an AEKey would have been observed in the network, it would never get
removed from the cached stacks. This causes performance to degrade over
time if many different items are transiting through the network.

The fix is simply the extra `removeEmptySubmaps` call. In this PR, I
also removed the backbuffer because it looks pointless in the way it is
currently implemented.