Avoid Object2LongMap.Entry allocations (#8362)
Small tweaks to reduce the number of entries that get allocated on each
server tick when updating the cached stacks.

A number of remaining allocations here appear to be from
https://github.com/AppliedEnergistics/Applied-Energistics-2/blob/126ff28d93ffc717c9dda8a40503660165e7a4de/src/main/java/appeng/api/stacks/VariantCounter.java#L95.
Ideally that should use `Object2LongMaps.fastIterator`, but I am not
sure if doing so is safe, since I'm not familiar with the rest of the
code.

---------

Co-authored-by: shartte <shartte@users.noreply.github.com>

(cherry picked from commit 1853dacc6479fb50fee9f628177f42c251d3153d)