Fix #6743: Shift-click insertion into the pattern access terminal (#8563)
Patterns are inserted into the pattern containers that are shown on the
client side, search bar filtering included. If there are multiple
groups, the pattern cannot be moved.

For crafting patterns, shift-clicking will typically put them in a
molecular assembler. For processing patterns, search for the name of the
target machine such that only one group of pattern containers is shown.