Balance annihilation plane enchantment energy cost (#8590)
The current implementation had two problems:

Enchanting an annihilation plane with any level of efficiency (and
nothing else) would reduce the required energy to zero due to
multiplying with the 0 in levelSum directly.

Enchanting an annihilation plane with unbreaking would increase the
energy cost by a factor of 8 per level of unbreaking, but only leading
to an average saving of 50% (lv1), 33% (lv2), or 25% (lv3), effectively
increasing the average cost rather than reducing it.

Fixes #8188

---------

Co-authored-by: Sebastian Hartte <shartte@users.noreply.github.com>